Add apps-pages gallery quality bar
