+++ src/agilab/apps-pages/view_relay_resilience/README.md
@@ -1,5 +1,10 @@
 # view_relay_resilience
 
+![view_relay_resilience preview](../../../../docs/source/_static/apps-pages-gallery/view_relay_resilience.svg)
+
+Package: `agi-page-relay-health`
+
+
 Streamlit analysis page for relay queue telemetry exported by any app that
 writes the AGILAB relay-resilience artifact contract.
